pyproject.toml
[project]
name = "dokdeploy"
version = "2.0.0"
description = "Deploy applications to Dokploy from the command line or GitHub Actions"
readme = "readme.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Taras Yarema", email = "taras@yarema.dev" }
]
keywords = ["dokploy", "deployment", "devops", "cli", "github-actions"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
dokdeploy = "src.cli:main"

[project.urls]
Homepage = "https://github.com/tarasyarema/dokploy-deploy-action"
Repository = "https://github.com/tarasyarema/dokploy-deploy-action"
Issues = "https://github.com/tarasyarema/dokploy-deploy-action/issues"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src"]

[tool.setuptools.package-data]
src = ["*.py"]
